@@ -5,8 +5,7 @@
     "type": "library",
     "authors": [
         {
-            "name": "Fredrik Liljegren",
-            "email": "fredrik.liljegren@textalk.se"
+            "name": "Fredrik Liljegren"
         },
         {
             "name": "Sören Jensen",
@@ -24,11 +23,12 @@
         }
     },
     "require": {
-        "php": "^5.4|^7.0"
+        "php": "^7.1",
+        "psr/log": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.1",
-        "php-coveralls/php-coveralls": "^0.7",
+        "phpunit/phpunit": "^7.0|^8.0|^9.0",
+        "php-coveralls/php-coveralls": "^2.0",
         "squizlabs/php_codesniffer": "^3.5"
     }
 }
