@@ -5,17 +5,21 @@
  * Note that this test is performed by mocking socket/stream calls.
  */
 
+declare(strict_types=1);
+
 namespace WebSocket;
 
-class ClientTest extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class ClientTest extends TestCase
 {
 
-    public function setUp()
+    public function setUp(): void
     {
         error_reporting(-1);
     }
 
-    public function testClientMasked()
+    public function testClientMasked(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -46,7 +50,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testDestruct()
+    public function testDestruct(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -56,7 +60,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         MockSocket::initialize('client.destruct', $this);
     }
 
-    public function testClienExtendedUrl()
+    public function testClienExtendedUrl(): void
     {
         MockSocket::initialize('client.connect-extended', $this);
         $client = new Client('ws://localhost:8000/my/mock/path?my_query=yes#my_fragment');
@@ -64,7 +68,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testClientWithTimeout()
+    public function testClientWithTimeout(): void
     {
         MockSocket::initialize('client.connect-timeout', $this);
         $client = new Client('ws://localhost:8000/my/mock/path', ['timeout' => 300]);
@@ -72,7 +76,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testClientWithContext()
+    public function testClientWithContext(): void
     {
         MockSocket::initialize('client.connect-context', $this);
         $client = new Client('ws://localhost:8000/my/mock/path', ['context' => '@mock-stream-context']);
@@ -80,7 +84,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testClientAuthed()
+    public function testClientAuthed(): void
     {
         MockSocket::initialize('client.connect-authed', $this);
         $client = new Client('wss://usename:password@localhost:8000/my/mock/path');
@@ -88,7 +92,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testWithHeaders()
+    public function testWithHeaders(): void
     {
         MockSocket::initialize('client.connect-headers', $this);
         $client = new Client('ws://localhost:8000/my/mock/path', [
@@ -99,7 +103,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testPayload128()
+    public function testPayload128(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -115,7 +119,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testPayload65536()
+    public function testPayload65536(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -133,7 +137,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(65540, $client->getFragmentSize());
     }
 
-    public function testMultiFragment()
+    public function testMultiFragment(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -149,7 +153,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(8, $client->getFragmentSize());
     }
 
-    public function testPingPong()
+    public function testPingPong(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -173,7 +177,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('ping', $client->getLastOpcode());
     }
 
-    public function testRemoteClose()
+    public function testRemoteClose(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -191,7 +195,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testSetTimeout()
+    public function testSetTimeout(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -204,7 +208,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testReconnect()
+    public function testReconnect(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
@@ -226,7 +230,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    public function testPersistentConnection()
+    public function testPersistentConnection(): void
     {
         MockSocket::initialize('client.connect-persistent', $this);
         $client = new Client('ws://localhost:8000/my/mock/path', ['persistent' => true]);
@@ -234,156 +238,132 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(MockSocket::isEmpty());
     }
 
-    /**
-     * @expectedException        WebSocket\BadUriException
-     * @expectedExceptionMessage Url should have scheme ws or wss
-     */
-    public function testBadScheme()
+    public function testBadScheme(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('bad://localhost:8000/my/mock/path');
+        $this->expectException('WebSocket\BadUriException');
+        $this->expectExceptionMessage('Url should have scheme ws or wss');
         $client->send('Connect');
     }
 
-    /**
-     * @expectedException        WebSocket\BadUriException
-     * @expectedExceptionMessage Invalid url 'this is not an url' provided.
-     */
-    public function testBadUrl()
+    public function testBadUrl(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('this is not an url');
+        $this->expectException('WebSocket\BadUriException');
+        $this->expectExceptionMessage('Invalid url \'this is not an url\' provided.');
         $client->send('Connect');
     }
 
-    /**
-     * @expectedException        InvalidArgumentException
-     * @expectedExceptionMessage Stream context in $options['context'] isn't a valid context
-     */
-    public function testBadStreamContext()
+    public function testBadStreamContext(): void
     {
         MockSocket::initialize('client.connect-bad-context', $this);
         $client = new Client('ws://localhost:8000/my/mock/path', ['context' => 'BAD']);
+        $this->expectException('InvalidArgumentException');
+        $this->expectExceptionMessage('Stream context in $options[\'context\'] isn\'t a valid context');
         $client->send('Connect');
     }
 
-    /**
-     * @expectedException        WebSocket\ConnectionException
-     * @expectedExceptionCode    0
-     * @expectedExceptionMessage Could not open socket to "localhost:8000"
-     */
-    public function testFailedConnection()
+    public function testFailedConnection(): void
     {
         MockSocket::initialize('client.connect-failed', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
+        $this->expectException('WebSocket\ConnectionException');
+        $this->expectExceptionCode(0);
+        $this->expectExceptionMessage('Could not open socket to "localhost:8000"');
         $client->send('Connect');
     }
 
-    /**
-     * @expectedException        WebSocket\ConnectionException
-     * @expectedExceptionCode    0
-     * @expectedExceptionMessage Connection to 'ws://localhost/my/mock/path' failed
-     */
-    public function testInvalidUpgrade()
+    public function testInvalidUpgrade(): void
     {
         MockSocket::initialize('client.connect-invalid-upgrade', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
+        $this->expectException('WebSocket\ConnectionException');
+        $this->expectExceptionCode(0);
+        $this->expectExceptionMessage('Connection to \'ws://localhost/my/mock/path\' failed');
         $client->send('Connect');
     }
 
-    /**
-     * @expectedException        WebSocket\ConnectionException
-     * @expectedExceptionCode    0
-     * @expectedExceptionMessage Server sent bad upgrade response
-     */
-    public function testInvalidKey()
+    public function testInvalidKey(): void
     {
         MockSocket::initialize('client.connect-invalid-key', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
+        $this->expectException('WebSocket\ConnectionException');
+        $this->expectExceptionCode(0);
+        $this->expectExceptionMessage('Server sent bad upgrade response');
         $client->send('Connect');
     }
 
-    /**
-     * @expectedException        WebSocket\BadOpcodeException
-     * @expectedExceptionMessage Bad opcode 'bad'.  Try 'text' or 'binary'.
-     */
-    public function testSendBadOpcode()
+    public function testSendBadOpcode(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
         $client->send('Connect');
 
         MockSocket::initialize('send-bad-opcode', $this);
+        $this->expectException('WebSocket\BadOpcodeException');
+        $this->expectExceptionMessage('Bad opcode \'bad\'.  Try \'text\' or \'binary\'.');
         $client->send('Bad Opcode', 'bad');
     }
 
-    /**
-     * @expectedException        WebSocket\ConnectionException
-     * @expectedExceptionCode    1026
-     * @expectedExceptionMessage Bad opcode in websocket frame: 12
-     */
-    public function testRecieveBadOpcode()
+    public function testRecieveBadOpcode(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
         $client->send('Connect');
         MockSocket::initialize('receive-bad-opcode', $this);
+        $this->expectException('WebSocket\ConnectionException');
+        $this->expectExceptionCode(1026);
+        $this->expectExceptionMessage('Bad opcode in websocket frame: 12');
         $message = $client->receive();
     }
 
-    /**
-     * @expectedException        WebSocket\ConnectionException
-     * @expectedExceptionCode    1025
-     * @expectedExceptionMessage Could only write 18 out of 22 bytes.
-     */
-    public function testBrokenWrite()
+    public function testBrokenWrite(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
         $client->send('Connect');
         MockSocket::initialize('send-broken-write', $this);
+        $this->expectException('WebSocket\ConnectionException');
+        $this->expectExceptionCode(1025);
+        $this->expectExceptionMessage('Could only write 18 out of 22 bytes.');
         $client->send('Failing to write');
     }
 
-    /**
-     * @expectedException        WebSocket\TimeoutException
-     * @expectedExceptionCode    1024
-     * @expectedExceptionMessage Failed to write 22 bytes.
-     */
-    public function testFailedWrite()
+    public function testFailedWrite(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
         $client->send('Connect');
         MockSocket::initialize('send-failed-write', $this);
+        $this->expectException('WebSocket\TimeoutException');
+        $this->expectExceptionCode(1024);
+        $this->expectExceptionMessage('Failed to write 22 bytes.');
         $client->send('Failing to write');
     }
 
-    /**
-     * @expectedException        WebSocket\ConnectionException
-     * @expectedExceptionCode    1025
-     * @expectedExceptionMessage Broken frame, read 0 of stated 2 bytes.
-     */
-    public function testBrokenRead()
+    public function testBrokenRead(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
         $client->send('Connect');
         MockSocket::initialize('receive-broken-read', $this);
+        $this->expectException('WebSocket\ConnectionException');
+        $this->expectExceptionCode(1025);
+        $this->expectExceptionMessage('Broken frame, read 0 of stated 2 bytes.');
         $client->receive();
     }
 
-    /**
-     * @expectedException        WebSocket\TimeoutException
-     * @expectedExceptionCode    1024
-     * @expectedExceptionMessage Empty read; connection dead?
-     */
-    public function testEmptyRead()
+    public function testEmptyRead(): void
     {
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
         $client->send('Connect');
         MockSocket::initialize('receive-empty-read', $this);
+        $this->expectException('WebSocket\TimeoutException');
+        $this->expectExceptionCode(1024);
+        $this->expectExceptionMessage('Empty read; connection dead?');
         $client->receive();
     }
 }
