@@ -18,6 +18,7 @@ class Client extends Base
       'fragment_size' => 4096,
       'context'       => null,
       'headers'       => null,
+      'logger'        => null,
       'origin'        => null, // @deprecated
     ];
 
@@ -36,6 +37,7 @@ class Client extends Base
     {
         $this->options = array_merge(self::$default_options, $options);
         $this->socket_uri = $uri;
+        $this->setLogger($this->options['logger']);
     }
 
     public function __destruct()
@@ -53,9 +55,9 @@ class Client extends Base
     {
         $url_parts = parse_url($this->socket_uri);
         if (empty($url_parts) || empty($url_parts['scheme']) || empty($url_parts['host'])) {
-            throw new BadUriException(
-                "Invalid url '$this->socket_uri' provided."
-            );
+            $error = "Invalid url '{$this->socket_uri}' provided.";
+            $this->logger->error($error);
+            throw new BadUriException($error);
         }
         $scheme    = $url_parts['scheme'];
         $host      = $url_parts['host'];
@@ -75,9 +77,9 @@ class Client extends Base
         }
 
         if (!in_array($scheme, array('ws', 'wss'))) {
-            throw new BadUriException(
-                "Url should have scheme ws or wss, not '$scheme' from URI '$this->socket_uri' ."
-            );
+            $error = "Url should have scheme ws or wss, not '{$scheme}' from URI '{$this->socket_uri}'.";
+            $this->logger->error($error);
+            throw new BadUriException($error);
         }
 
         $host_uri = ($scheme === 'wss' ? 'ssl' : 'tcp') . '://' . $host;
@@ -88,9 +90,9 @@ class Client extends Base
             if (@get_resource_type($this->options['context']) === 'stream-context') {
                 $context = $this->options['context'];
             } else {
-                throw new \InvalidArgumentException(
-                    "Stream context in \$options['context'] isn't a valid context"
-                );
+                $error = "Stream context in \$options['context'] isn't a valid context.";
+                $this->logger->error($error);
+                throw new \InvalidArgumentException($error);
             }
         } else {
             $context = stream_context_create();
@@ -110,9 +112,9 @@ class Client extends Base
         );
 
         if (!$this->isConnected()) {
-            throw new ConnectionException(
-                "Could not open socket to \"$host:$port\": $errstr ($errno)."
-            );
+            $error = "Could not open socket to \"{$host}:{$port}\": {$errstr} ({$errno}).";
+            $this->logger->error($error);
+            throw new ConnectionException($error);
         }
 
         // Set timeout on the stream as well.
@@ -165,13 +167,13 @@ class Client extends Base
 
         /// @todo Handle version switching
 
+        $address = "{$scheme}://{$host}{$path_with_query}";
+
         // Validate response.
         if (!preg_match('#Sec-WebSocket-Accept:\s(.*)$#mUi', $response, $matches)) {
-            $address = $scheme . '://' . $host . $path_with_query;
-            throw new ConnectionException(
-                "Connection to '{$address}' failed: Server sent invalid upgrade response:\n"
-                . $response
-            );
+            $error = "Connection to '{$address}' failed: Server sent invalid upgrade response: {$response}";
+            $this->logger->error($error);
+            throw new ConnectionException($error);
         }
 
         $keyAccept = trim($matches[1]);
@@ -179,8 +181,12 @@ class Client extends Base
             = base64_encode(pack('H*', sha1($key . '258EAFA5-E914-47DA-95CA-C5AB0DC85B11')));
 
         if ($keyAccept !== $expectedResonse) {
-            throw new ConnectionException('Server sent bad upgrade response.');
+            $error = 'Server sent bad upgrade response.';
+            $this->logger->error($error);
+            throw new ConnectionException($error);
         }
+
+        $this->logger->info("Client connected to to {$address}");
     }
 
     /**
