@@ -5,7 +5,7 @@ namespace WebSocket;
 class ConnectionException extends Exception
 {
     // Native codes in interval 0-106
-    const TIMED_OUT = 1024;
-    const EOF = 1025;
-    const BAD_OPCODE = 1026;
+    public const TIMED_OUT = 1024;
+    public const EOF = 1025;
+    public const BAD_OPCODE = 1026;
 }
